Remove broken functionality for removing context items
