+++ SolidDna/CADBooster.SolidDna/SolidWorks/CommandManager/Item/ICommandCreatable.cs
@@ -5,7 +5,6 @@ namespace CADBooster.SolidDna.SolidWorks.CommandManager.Item
 {
     public interface ICommandCreatable
     {
-        int UserId { get; }
         string Name { get; }
 
         IEnumerable<ICommandCreated> Create(string path = "");
