+++ SolidDna/CADBooster.SolidDna/SolidWorks/CommandManager/Item/ICommandCreated.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace CADBooster.SolidDna
+{
+    public interface ICommandCreated : IDisposable
+    {
+        string Name { get; }
+    }
+}
